# api/swagger.yaml
swagger: '2.0'
info:
  version: 1.0.2
  title: Smart House
  description: The API for the Smart House IOT project for embedded systems course
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  ApiKeyAuth:
    type: apiKey
    in: header
    name: SmartHouseSession
security:
  - ApiKeyAuth: []
paths:
...
  /lights/{lightID}/{state}:
    put:
      tags:
      - Lights
      description: turns the light on or off
      operationId: setLightState
      parameters:
      - name: lightID
        in: path
        required: true
        type: string
      - name: state
        in: path
        required: true
        type: string
        enum:
        - on
        - off
      responses:
        200:
          description: Successful light state change
          schema:
            $ref: '#/definitions/StatusResponse'
...
  /music/{state}:
    put:
        tags:
        - Music
        description: turns the music on or off
        operationId: setMusicState
        parameters:
        - name: state
          in: path
          required: true
          type: string
          enum:
          - on
          - off
        responses:
          200:
            description: Successful music toggle
            schema:
              $ref: '#/definitions/MusicPlayerStatus'
  /music/play:
    put:
      tags:
      - Music
      description: play track based on query
      operationId: playTrack
      parameters:
      - name: trackId
        in: query
        required: true
        type: integer
      responses:
        200:
          description: Successful play request
          schema:
            $ref: '#/definitions/StatusResponse'
  /settings/home/:
    put:
        tags:
        - Settings
        description: set global settings
        operationId: setHomeSettings
        parameters:
        - name: settings
          in: body
          required: true
          schema:
            $ref: '#/definitions/Setting'
        responses:
          200:
            description: Current Global Settings
            schema:
              $ref: '#/definitions/StatusResponse'
definitions:
...
  MusicPlayerStatus:
    type: object
    required:
      - status
    properties:
      status:
        type: boolean
      track:
        $ref: '#/definitions/Track'
  Track:
    type: object
    required:
      - id
      - name
    properties:
      id:
        type: integer
      name:
        type: string
    example:
      id: 42
      name: "Rick Astley - Never Gonna Give You Up"
...
  Setting:
    type: object
    properties:
      automatic:
        type: boolean
      threshold:
        type: number
    example:
      automatic: true
      threshold: 1.5
  StatusResponse:
    type: object
    properties:
      message:
        type: string
        example: 'OK'
        
# Added by API Auto Mocking Plugin
host: virtserver.swaggerhub.com
basePath: /Evilong/SmartHouse/1.0.2
schemes:
 - https